requirements.txt
python-telegram-bot==20.7
gspread==6.1.2
oauth2client==4.1.3